requirements.txt
backports.shutil-get-terminal-size==1.0.0
click==6.7
decorator==4.2.1
enum34==1.1.6
Flask==0.12.2
FreeSWITCH-ESL-Python==1.2
ipdb==0.11
ipython==5.6.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uWSGI==2.0.17
wcwidth==0.1.7
Werkzeug==0.14.1
